Refactor: Move classifier classes out of Service folder
