+++ lib/Classifiers/Images/FacesClassifier.php
@@ -17,10 +17,9 @@ use Symfony\Component\Process\Exception\ProcessTimedOutException;
 use Symfony\Component\Process\Exception\RuntimeException;
 use Symfony\Component\Process\Process;
 
-class ClassifyImagenetService {
-    public const IMAGE_TIMEOUT = 480; // seconds
-    public const IMAGE_PUREJS_TIMEOUT = 600; // seconds
-    public const MODEL_DOWNLOAD_TIMEOUT = 180; // seconds
+class FacesClassifier {
+    public const IMAGE_TIMEOUT = 120; // seconds
+    public const IMAGE_PUREJS_TIMEOUT = 360; // seconds
 
     /**
      * @var LoggerInterface
@@ -46,20 +45,17 @@ class ClassifyImagenetService {
      * @return void
      * @throws \OCP\Files\NotFoundException
      */
-    public function classify(array $files): void {
-        if ($this->config->getAppValue('recognize', 'imagenet.enabled', 'false') !== 'true') {
-            return;
-        }
-
+    public function classify(array $faces, array $files): void {
         $paths = array_map(static function($file) {
             return $file->getStorage()->getLocalFile($file->getInternalPath());
         }, $files);
 
+        $this->logger->debug('Reference faces '.var_export($faces, true));
         $this->logger->debug('Classifying '.var_export($paths, true));
 
         $command = [
             $this->config->getAppValue('recognize', 'node_binary'),
-            dirname(__DIR__, 2) . '/src/classifier_imagenet.js',
+            dirname(__DIR__, 2) . '/src/classifier_faces.js',
             '-'
         ];
 
@@ -70,11 +66,11 @@ class ClassifyImagenetService {
         }
         if ($this->config->getAppValue('recognize', 'tensorflow.purejs', 'false') === 'true') {
             $proc->setEnv(['RECOGNIZE_PUREJS' => 'true']);
-            $proc->setTimeout(count($paths) * self::IMAGE_PUREJS_TIMEOUT + self::MODEL_DOWNLOAD_TIMEOUT);
+            $proc->setTimeout(count($paths) * self::IMAGE_PUREJS_TIMEOUT);
         }else{
-            $proc->setTimeout(count($paths) * self::IMAGE_TIMEOUT + self::MODEL_DOWNLOAD_TIMEOUT);
+            $proc->setTimeout(count($paths) * self::IMAGE_TIMEOUT);
         }
-        $proc->setInput(implode("\n", $paths));
+        $proc->setInput(json_encode($faces)."\n\n".implode("\n", $paths));
         try {
             $proc->start();
